metabase: separate grant and revoke
